Move HostList to pages
