+++ assets/js/common/SapSystemLink/index.js
@@ -1,0 +1,3 @@
+import SapSystemLink from './SapSystemLink';
+
+export default SapSystemLink;
