--- assets/js/pages/HostList/index.js
@@ -1,3 +1,0 @@
-import HostList from './HostList';
-
-export default HostList;
