+++ assets/js/pages/HostsList/index.js
@@ -1,0 +1,3 @@
+import HostsList from './HostsList';
+
+export default HostsList;
